Rename to IngestError to support SubmissionError as a sub-class of ImportError, as original functionality.
